+++ TrainStation/docs/release-notes.md
@@ -5,6 +5,7 @@
 - Removed the support for Content Patcher destinations added in 2.0.0.  
   _This needs a bit more work, and will be released as a separate mod instead. However, all the other improvements and
   stability fixes from the 2.0.0 rewrite still apply._
+- Fixed train sounds looping forever in some cases.
 - Fixed content packs which provide both boat and train stops sometimes warping to the wrong stop.
 - Improved translations. Thanks to BlackRosePetals (updated Chinese)!